Add files via upload
This is basic excel project where we created Bike Sales Dashboard. 
In this we created a copy of the raw data and then first removed duplicates, removed unwanted columns if any and also standardized the data. 
We created a Pivot table. 
We created a dashboard on the basis of the pivot table we created for the data.